feat: 예약된 시간을 삭제하려는 경우 400 BadRequest 를 반환하는 기능 구현

+++ src/main/java/roomescape/service/reservationtime/ReservationTimeDeleteService.java
@@ -1,15 +1,18 @@
 package roomescape.service.reservationtime;
 
 import org.springframework.stereotype.Service;
+import roomescape.repository.ReservationRepository;
 import roomescape.repository.ReservationTimeRepository;
 
 @Service
 public class ReservationTimeDeleteService {
 
     private final ReservationTimeRepository reservationTimeRepository;
+    private final ReservationRepository reservationRepository;
 
-    public ReservationTimeDeleteService(ReservationTimeRepository reservationTimeRepository) {
+    public ReservationTimeDeleteService(ReservationTimeRepository reservationTimeRepository, ReservationRepository reservationRepository) {
         this.reservationTimeRepository = reservationTimeRepository;
+        this.reservationRepository = reservationRepository;
     }
 
     public void deleteReservationTime(Long id) {
@@ -19,6 +22,9 @@ public class ReservationTimeDeleteService {
                 .findFirst()
                 .orElseThrow(() -> new IllegalArgumentException("존재하지 않는 예약 아이디 입니다."));
 
+        if(reservationRepository.existByReservationTimeId(id)) {
+            throw new IllegalArgumentException("이미 예약중인 시간은 삭제할 수 없습니다.");
+        }
         reservationTimeRepository.deleteById(id);
     }
 }
